feat: first implementation of data provider
In order to create a candidate's dataset, it's necessary to collect/mining data about the candidates and then provide the data. This algorithm was developed only to provide the candidate's data to another code and then create the dataset. The algorithm is not completed yet, it's possible to have some changes.

See also: #4, #3