feat: implementation of a voter's provider
To create the second dataset, instead of count words from candidate's proposal, this time the data will be voter's rating about some candidate. The rating is calculated by giving rating to some areas and then pick up the mean or median. This code was created to optimize this step by creating it automatically.