feat: collaborative filtering of voters' provider
After voters' provider implementation, this code was created to implement a collaborative filtering for this dataset. The result is a recommendation of a candidate to some voter by seeing his nearest neighbors and using manhattan distance. An analysis will be made to verify the efficiency of the algorithm.
See also: #5